# core/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_alpkeskin_rota_core_internal_models.BulkCreateProxyRequest:
    properties:
      proxies:
        items:
          $ref: '#/definitions/github_com_alpkeskin_rota_core_internal_models.CreateProxyRequest'
        minItems: 1
        type: array
    required:
    - proxies
    type: object
  github_com_alpkeskin_rota_core_internal_models.BulkDeleteProxyRequest:
    properties:
      ids:
        items:
          type: integer
        minItems: 1
        type: array
    required:
    - ids
    type: object
...
  github_com_alpkeskin_rota_core_internal_models.CreateProxyRequest:
    properties:
      address:
        type: string
      password:
        type: string
      protocol:
        enum:
        - http
        - https
        - socks4
        - socks4a
        - socks5
        type: string
      username:
        type: string
    required:
    - address
    - protocol
    type: object
...
  github_com_alpkeskin_rota_core_internal_models.ErrorResponse:
    properties:
      details: {}
      error:
        type: string
    type: object
...
  github_com_alpkeskin_rota_core_internal_models.PaginationMeta:
    properties:
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
  github_com_alpkeskin_rota_core_internal_models.Proxy:
    properties:
      address:
        type: string
      avg_response_time:
        type: integer
      created_at:
        type: string
      id:
        type: integer
      last_check:
        type: string
      protocol:
        type: string
      requests:
        type: integer
      status:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
  github_com_alpkeskin_rota_core_internal_models.ProxyListResponse:
    properties:
      pagination:
        $ref: '#/definitions/github_com_alpkeskin_rota_core_internal_models.PaginationMeta'
      proxies:
        items:
          $ref: '#/definitions/github_com_alpkeskin_rota_core_internal_models.ProxyWithStats'
        type: array
    type: object
  github_com_alpkeskin_rota_core_internal_models.ProxyTestResult:
    properties:
      address:
        type: string
      error:
        type: string
      id:
        type: integer
      response_time:
        type: integer
      status:
        type: string
      tested_at:
        type: string
    type: object
  github_com_alpkeskin_rota_core_internal_models.ProxyWithStats:
    properties:
      address:
        type: string
      avg_response_time:
        type: integer
      created_at:
        type: string
      id:
        type: integer
      last_check:
        type: string
      protocol:
        type: string
      requests:
        type: integer
      status:
        type: string
      success_rate:
        type: number
      updated_at:
        type: string
      username:
        type: string
    type: object
...
  github_com_alpkeskin_rota_core_internal_models.UpdateProxyRequest:
    properties:
      address:
        type: string
      password:
        type: string
      protocol:
        enum:
        - http
        - https
        - socks4
        - socks4a
        - socks5
        type: string
      username:
        type: string
    type: object
...
host: localhost:8001
info:
  contact:
    name: API Support
    url: https://github.com/alpkeskin/rota
  description: |-
    A high-performance proxy rotation server with health monitoring and intelligent routing
    Provides comprehensive API for managing proxy servers, monitoring their health,
    and configuring rotation strategies.
  license:
    name: LICENSE
    url: https://github.com/alpkeskin/rota/blob/main/LICENSE
  title: Rota Proxy API
  version: 1.0.0
paths:
...
  /proxies:
    get:
      description: Get paginated list of proxies with optional filters
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      - description: Search term
        in: query
        name: search
        type: string
      - description: Filter by status
        in: query
        name: status
        type: string
      - description: Filter by protocol
        in: query
        name: protocol
        type: string
      - description: Sort field
        in: query
        name: sort
        type: string
      - description: Sort order (asc/desc)
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of proxies
          schema:
            $ref: '#/definitions/github_com_alpkeskin_rota_core_internal_models.ProxyListResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_alpkeskin_rota_core_internal_models.ErrorResponse'
      summary: List proxies
      tags:
      - proxies
    post:
      consumes:
      - application/json
      description: Create a new proxy server
      parameters:
      - description: Proxy details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_alpkeskin_rota_core_internal_models.CreateProxyRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created proxy
          schema:
            $ref: '#/definitions/github_com_alpkeskin_rota_core_internal_models.Proxy'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_alpkeskin_rota_core_internal_models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_alpkeskin_rota_core_internal_models.ErrorResponse'
      summary: Create proxy
      tags:
      - proxies
  /proxies/{id}:
    delete:
      description: Delete a proxy server by ID
      parameters:
      - description: Proxy ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: Successfully deleted
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_alpkeskin_rota_core_internal_models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_alpkeskin_rota_core_internal_models.ErrorResponse'
      summary: Delete proxy
      tags:
      - proxies
    put:
      consumes:
      - application/json
      description: Update an existing proxy server
      parameters:
      - description: Proxy ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated proxy details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_alpkeskin_rota_core_internal_models.UpdateProxyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Updated proxy
          schema:
            $ref: '#/definitions/github_com_alpkeskin_rota_core_internal_models.Proxy'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_alpkeskin_rota_core_internal_models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_alpkeskin_rota_core_internal_models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_alpkeskin_rota_core_internal_models.ErrorResponse'
      summary: Update proxy
      tags:
      - proxies
  /proxies/{id}/test:
    post:
      description: Test a proxy server's connectivity and performance
      parameters:
      - description: Proxy ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Test results
          schema:
            $ref: '#/definitions/github_com_alpkeskin_rota_core_internal_models.ProxyTestResult'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_alpkeskin_rota_core_internal_models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_alpkeskin_rota_core_internal_models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_alpkeskin_rota_core_internal_models.ErrorResponse'
      summary: Test proxy
      tags:
      - proxies
  /proxies/bulk:
    post:
      consumes:
      - application/json
      description: Create multiple proxy servers at once
      parameters:
      - description: List of proxies to create
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_alpkeskin_rota_core_internal_models.BulkCreateProxyRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Creation results
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_alpkeskin_rota_core_internal_models.ErrorResponse'
      summary: Bulk create proxies
      tags:
      - proxies
  /proxies/bulk-delete:
    post:
      consumes:
      - application/json
      description: Delete multiple proxy servers at once
      parameters:
      - description: List of proxy IDs to delete
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_alpkeskin_rota_core_internal_models.BulkDeleteProxyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Deletion results
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_alpkeskin_rota_core_internal_models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_alpkeskin_rota_core_internal_models.ErrorResponse'
      summary: Bulk delete proxies
      tags:
      - proxies
  /proxies/export:
    get:
      description: Export proxy list in various formats (txt, json, csv)
      parameters:
      - default: txt
        description: Export format (txt/json/csv)
        in: query
        name: format
        type: string
      - description: Filter by status
        in: query
        name: status
        type: string
      produces:
      - text/plain
      - application/json
      - text/csv
      responses:
        "200":
          description: Exported file
          schema:
            type: file
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_alpkeskin_rota_core_internal_models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_alpkeskin_rota_core_internal_models.ErrorResponse'
      summary: Export proxies
      tags:
      - proxies
...
  /status:
    get:
      description: Get detailed system status including proxy and request statistics
      produces:
      - application/json
      responses:
        "200":
          description: Detailed status information
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_alpkeskin_rota_core_internal_models.ErrorResponse'
      summary: System status
      tags:
      - health
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"